feature: implement payment service and add custom query repository method

+++ src/main/java/com/demo/coffeeassessment/core/domain/Payment.java
@@ -1,3 +1,3 @@
 package com.demo.coffeeassessment.core.domain;
 
-public record Payment(String user, int amount) {}
+public record Payment(String user, double amount) {}
